backend/requirements.txt
asgiref==3.8.1
boto3==1.26.77
Django==4.2.11
django-ipware==4.0.2
django-debug-toolbar==4.1.0
djangorestframework==3.15.1
environs==9.5.0
gunicorn==22.0.0
psycopg==3.1.18
Faker==16.1.0
pycryptodome==3.19.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
